Add Page Report for private event
