+++ src/modules/BettingReport/BettingReport.scss
@@ -1,0 +1,6 @@
+.form-admin{
+  .form-check{
+    display: inline-block;
+    width: 120px;
+  }
+}
