+++ src/modules/BettingReport/index.js
@@ -1,0 +1,3 @@
+import BettingReport from './BettingReport';
+
+export default BettingReport;
