+++ src/pages/Admin/Admin.scss
@@ -1,6 +1,1 @@
-.form-admin{
-  .form-check{
-    display: inline-block;
-    width: 120px;
-  }
-}
+
